test/db-console-standalone/pyproject.toml
[project]
name = "db-console-standalone"
version = "0.1.0"
description = "PostgreSQL Web 控制台（Flask + 静态 HTML/JS）"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.0.0",
    "flask-cors>=4.0.0",
]

[tool.uv]
package = false
